Rename `lib/zuul.js` to `lib/airtap.js`
